+++ src/components/ProductSwitcher/ProductSwitcher.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
 import { cnCreate } from '../../utils/cn';
+import modernizr from '../../utils/modernizr';
 import './ProductSwitcher.less';
 
 interface IItem {
@@ -26,6 +27,15 @@ interface IProductSwitcherState {
     /** Current value */
     currentValue: string;
     currentIndex: number;
+    isPointerPressed: boolean;
+    isPressedPointerLeaveBlock: boolean;
+}
+
+interface INearPoint {
+    item: number;
+    coord: number;
+    value: string;
+    title: string;
 }
 
 const cn = cnCreate('mfui-product-switcher');
@@ -41,11 +51,13 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
         startIndex: 0,
     };
 
-    itemNodes: object;
+    labelNodes: object;
     rootNode: HTMLElement | null;
     pointerNode: HTMLElement | null;
     colorRowNode: HTMLElement | null;
     timer: number;
+    rowItemsInfo: INearPoint[];
+    switcherStep: number;
 
     constructor(props: IProductSwitcherProps) {
         super(props);
@@ -53,18 +65,23 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
         const { items, startIndex = 0 } = props;
         const safeStartIndex: number = this.getSafeStartIndex(startIndex, items);
 
-        this.itemNodes = {};
-
-        this.state = { currentValue: items[safeStartIndex].value, currentIndex: safeStartIndex };
+        this.labelNodes = {};
+        this.state = {
+            currentValue: items[safeStartIndex].value,
+            currentIndex: safeStartIndex,
+            isPointerPressed: false,
+            isPressedPointerLeaveBlock: false,
+        };
     }
 
     componentDidMount() {
-        window.addEventListener('resize', this.handleWindowSize, true);
+        this.rowItemsInfo = this.getRowItemsInfo();
         this.movePointer(this.state.currentValue);
+        this.setHandlers();
     }
 
     componentWillUnmount() {
-        window.removeEventListener('resize', this.handleWindowSize, true);
+        this.removeHandlers();
     }
 
     componentDidUpdate(prevProps: IProductSwitcherProps) {
@@ -87,6 +104,8 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
     }
 
     handleWindowSize = () => {
+        this.rowItemsInfo = this.getRowItemsInfo();
+
         clearTimeout(this.timer);
 
         this.timer = window.setTimeout(() => {
@@ -107,10 +126,282 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
         this.movePointer(value);
     }
 
+    setHandlers = () => {
+        window.addEventListener('resize', this.handleWindowSize, true);
+
+        if (!this.pointerNode || !this.rootNode) {
+            return;
+        }
+
+        const pointerNode = this.pointerNode;
+        const rootNode = this.rootNode;
+
+        if (modernizr.touchevents) {
+            pointerNode.addEventListener('touchstart', this.setPointerState);
+            rootNode.addEventListener('touchmove', this.handleTouchMove);
+            rootNode.addEventListener('touchend', this.handleTouchEnd);
+
+            return;
+        }
+
+        pointerNode.addEventListener('mousedown', this.setPointerState);
+        rootNode.addEventListener('mousemove', this.handleMouseMove);
+        rootNode.addEventListener('mouseleave', this.handleMouseLeave);
+        rootNode.addEventListener('mouseup', this.handleMouseUp);
+        document.body.addEventListener('mouseup', this.handleBodyMouseUp);
+    }
+
+    removeHandlers = () => {
+        window.removeEventListener('resize', this.handleWindowSize, true);
+
+        if (!this.pointerNode || !this.rootNode) {
+            return;
+        }
+
+        const pointerNode = this.pointerNode;
+        const rootNode = this.rootNode;
+
+        if (modernizr.touchevents) {
+            pointerNode.removeEventListener('touchstart', this.setPointerState);
+            rootNode.removeEventListener('touchmove', this.handleTouchMove);
+            rootNode.removeEventListener('touchend', this.handleTouchEnd);
+
+            return;
+        }
+
+        pointerNode.removeEventListener('mousedown', this.setPointerState);
+        rootNode.removeEventListener('mousemove', this.handleMouseMove);
+        rootNode.removeEventListener('mouseleave', this.handleMouseLeave);
+        rootNode.removeEventListener('mouseup', this.handleMouseUp);
+        document.body.removeEventListener('mouseup', this.handleBodyMouseUp);
+    }
+
+    setPointerState = (e: Event) => {
+        e.stopPropagation();
+
+        this.setState(prevState => ({
+            isPointerPressed: !prevState.isPointerPressed,
+        }));
+    }
+
+    handleTouchMove = (e: TouchEvent) => {
+        const { isPointerPressed } = this.state;
+
+        if (!this.rootNode || !isPointerPressed) {
+            return;
+        }
+
+        e.stopPropagation();
+
+        const { top, bottom } = this.getRangeWrapperCoords(this.rootNode);
+        const eventXCoord = e.changedTouches[0].clientX;
+        const eventYCoord = e.changedTouches[0].clientY;
+
+        if (eventYCoord < top || eventYCoord > bottom) {
+            this.handleTouchEnd(e);
+
+            return;
+        }
+
+        this.moveSwitcher(eventXCoord);
+    }
+
+    handleMouseMove = (e: MouseEvent) => {
+        const { isPointerPressed } = this.state;
+
+        if (!isPointerPressed) {
+            return;
+        }
+
+        e.stopPropagation();
+
+        this.moveSwitcher(e.clientX);
+    }
+
+    handleMouseLeave = (e: MouseEvent) => {
+        const { isPointerPressed } = this.state;
+
+        if (!isPointerPressed || !this.rootNode) {
+            return;
+        }
+
+        e.stopPropagation();
+
+        const {
+            left: startPoint = 0,
+            right: endPoint = 0,
+            top,
+            bottom,
+        } = this.getRangeWrapperCoords(this.rootNode);
+        const eventYCoord = e.clientY;
+        const eventXCoord = e.clientX;
+
+        if (eventYCoord < top || eventYCoord > bottom) {
+            this.handleMouseUp(e);
+
+            return;
+        }
+
+        if (eventXCoord < startPoint || eventXCoord > endPoint) {
+            this.setState({
+                isPressedPointerLeaveBlock: true,
+            });
+        }
+    }
+
+    handleMouseUp = (e: MouseEvent) => {
+        const { isPointerPressed, isPressedPointerLeaveBlock } = this.state;
+
+        if (!isPointerPressed) {
+            return;
+        }
+
+        if (isPressedPointerLeaveBlock) {
+            this.setState({
+                isPressedPointerLeaveBlock: false,
+            });
+        }
+
+        this.handleEndSwitchActions(e.clientX);
+    }
+
+    handleBodyMouseUp = (e: MouseEvent) => {
+        const { isPressedPointerLeaveBlock } = this.state;
+
+        if (!isPressedPointerLeaveBlock) {
+            return;
+        }
+
+        this.handleMouseUp(e);
+    }
+
+    handleTouchEnd = (e: TouchEvent) => {
+        const { isPointerPressed } = this.state;
+
+        if (!isPointerPressed) {
+            return;
+        }
+
+        this.handleEndSwitchActions(e.changedTouches[0].clientX);
+    }
+
+    moveSwitcher = (eventXCoord: number) => {
+        if (!this.pointerNode || !this.colorRowNode || !this.rootNode) {
+            return;
+        }
+
+        const {
+            left: startPoint = 0,
+            right: endPoint = 0,
+            width = 0,
+        } = this.getRangeWrapperCoords(this.rootNode);
+        const pointHalfWidth = this.pointerNode.offsetWidth / 2 || 0;
+
+        switch (true) {
+            case eventXCoord < startPoint + pointHalfWidth:
+                this.pointerNode.style.transform = 'translateX(0px)';
+                this.colorRowNode.style.width = '0px';
+                break;
+            case eventXCoord >= endPoint - pointHalfWidth:
+                this.pointerNode.style.transform = `translateX(${width - pointHalfWidth * 2}px)`;
+                this.colorRowNode.style.width = `${width - pointHalfWidth}px`;
+                break;
+            default:
+                this.pointerNode.style.transform = `translateX(${eventXCoord - startPoint - pointHalfWidth}px)`;
+                this.colorRowNode.style.width = `${eventXCoord - startPoint}px`;
+        }
+    }
+
+    handleEndSwitchActions = (eventXCoord: number) => {
+        if (!this.rootNode) {
+            return;
+        }
+
+        const { left: startPoint = 0 } = this.getRangeWrapperCoords(this.rootNode);
+        const outRowPoint = eventXCoord - startPoint;
+        const [nearPoint] = this.getNearPoint(outRowPoint);
+
+        this.setState(prevState => ({
+            currentValue: nearPoint.value,
+            currentIndex: nearPoint.item,
+            isPointerPressed: !prevState.isPointerPressed,
+        }));
+
+        this.movePointer(nearPoint.value);
+    }
+
+    getNearPoint = (outRowPoint: number) => {
+        return this.rowItemsInfo.filter((el: INearPoint, ind: number, arr: [INearPoint]) => {
+            const prevEl = arr[ind - 1];
+
+            if (ind === 0) {
+                return outRowPoint <= this.switcherStep / 2;
+            }
+
+            if (ind === arr.length - 1) {
+                return outRowPoint >= prevEl.coord + this.switcherStep / 2;
+            }
+
+            return outRowPoint >= el.coord - this.switcherStep / 2 && outRowPoint <= el.coord + this.switcherStep / 2;
+        });
+    }
+
+    getRangeWrapperCoords = (node: HTMLElement) => {
+        return node && node.getBoundingClientRect();
+    }
+
+    getRowItemsInfo = () => {
+        if (!this.rootNode) {
+            return [];
+        }
+
+        const { items } = this.props;
+        const { width: rowWidth = 0 } = this.getRangeWrapperCoords(this.rootNode);
+        const itemsLength = items.length;
+
+        this.setSwitcherStep(rowWidth / (itemsLength - 1));
+
+        return items.map((item: IItem, index: number) => {
+            if (index === 0) {
+                return {
+                    ...item,
+                    item: 0,
+                    coord: 0,
+                };
+            }
+
+            if (index === itemsLength - 1) {
+                return {
+                    ...item,
+                    item: itemsLength - 1,
+                    coord: rowWidth,
+                };
+            }
+
+            return {
+                ...item,
+                item: index,
+                coord: Math.floor((rowWidth / (itemsLength - 1)) * index),
+            };
+        });
+    }
+
+    setSwitcherStep = (value: number) => {
+        this.switcherStep = value;
+    }
+
     movePointer(value: string) {
+        const { theme } = this.props;
+
+        if (theme === 'with-color-row') {
+            this.moveSwipePointer(value);
+
+            return;
+        }
+
         const rootOffsetLeft = this.rootNode ? this.rootNode.getBoundingClientRect().left : 0;
-        const hasLabelProp = Object.keys(this.itemNodes).length !== 0;
-        const targetLabelOffsetLeft = hasLabelProp ? this.itemNodes[value].getBoundingClientRect().left : 0;
+        const hasLabelProp = Object.keys(this.labelNodes).length !== 0;
+        const targetLabelOffsetLeft = hasLabelProp ? this.labelNodes[value].getBoundingClientRect().left : 0;
         const offsetValue = targetLabelOffsetLeft - rootOffsetLeft;
 
         if (this.pointerNode) {
@@ -122,12 +413,39 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
         }
     }
 
+    moveSwipePointer = (value: string) => {
+        if (!this.pointerNode || !this.colorRowNode) {
+            return;
+        }
+
+        const [checkedItem] = this.rowItemsInfo.filter((el: INearPoint ) => el.value === value);
+        let switchPointOffsetValue;
+        let switchRowWidth;
+
+        switch (!!checkedItem) {
+            case checkedItem.item === 0:
+                switchPointOffsetValue = 0;
+                switchRowWidth = 0;
+                break;
+            case checkedItem.item === this.rowItemsInfo.length - 1:
+                switchPointOffsetValue = checkedItem.coord - (this.pointerNode.offsetWidth);
+                switchRowWidth = checkedItem.coord - (this.pointerNode.offsetWidth / 2);
+                break;
+            default:
+                switchPointOffsetValue = checkedItem.coord - (this.pointerNode.offsetWidth / 2);
+                switchRowWidth = checkedItem.coord;
+        }
+
+        this.pointerNode.style.transform = `translateX(${switchPointOffsetValue}px)`;
+        this.colorRowNode.style.width = `${switchRowWidth}px`;
+    }
+
     getRootNode = (node: HTMLElement | null) => {
         this.rootNode = node;
     }
 
-    getItemNodes = (item: IItem) => (node: HTMLElement | null) =>  {
-        this.itemNodes[item.value] = node;
+    getLabelNodes = (value: string) => (node: HTMLElement | null) =>  {
+        this.labelNodes[value] = node;
     }
 
     getPointerNode = (node: HTMLElement | null) => {
@@ -146,23 +464,24 @@ class ProductSwitcher extends React.Component<IProductSwitcherProps, IProductSwi
                 <div className={cn('row')}>
                     {items.map((item, index) =>
                         <div
-                            className={cn('item', {
-                                active: item.value === this.state.currentValue,
-                                disabled: index < this.state.currentIndex,
-                            })}
+                            className={cn('item',
+                                {
+                                    active: item.value === this.state.currentValue,
+                                    disabled: index < this.state.currentIndex,
+                                }
+                            )}
                             key={item.title + item.value}
                             onClick={this.handleClickItem(item.value, index)}
-                            ref={this.getItemNodes(item)}
                         >
-                            <div className={cn('label')}>
+                            <div className={cn('label')} ref={this.getLabelNodes(item.value)}>
                                 {item.title}
                                 {item.unit && <span className={cn('label-unit')}>{item.unit}</span>}
                             </div>
                         </div>
                     )}
                 </div>
-                <div className={cn('pointer')} ref={this.getPointerNode}/>
-                <div className={cn('color-row')} ref={this.getColorRowNode}/>
+                <div className={cn('pointer')} ref={this.getPointerNode} />
+                <div className={cn('color-row')} ref={this.getColorRowNode} />
             </div>
         );
     }