+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - **Carousel**
     - fix carousel styles
 
+### Changed
+- **ProductSwitcher**
+    - add swipe functionality
+
 ## Release 1.3.33
 ### Changed
 - **package.json**